Limits to coinc finder and bug fix multi file

+++ ImageReconstruction/CoInicidenceFinder/EventEntry.cpp
@@ -1,6 +1,6 @@
 #include "EventEntry.h"
 
-EventEntry::EventEntry(unsigned int timeStamp, unsigned int energy, int detector)
+EventEntry::EventEntry(unsigned int timeStamp, double energy, int detector)
 {
 	_TimeStamp = timeStamp;
